# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IoT 物联网平台后端 API（Go 重构版）",
        "title": "IoT Platform Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@iot-platform.local.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.7.0"
    },
    "host": "localhost:8182",
    "basePath": "/",
    "paths": {
        "/api/devices": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "查询用户设备列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ent.Device"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "注册设备",
                "parameters": [
                    {
                        "description": "设备注册请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DeviceParameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.DeviceRegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
        "/api/devices/{deviceId}": {
            "get": {
                "description": "设备元信息 + 在线状态 + sensors（定义+latest）/ actuators（定义）物模型数组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "获取设备详情（物模型视图：定义 + 最近遥测 + 执行器）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID（6位hex）",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
...
        "/api/devices/{deviceId}/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
...
        "/api/devices/{deviceId}/login": {
            "get": {
                "description": "使用设备6位hex ID认证（路径参数），无需额外Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "获取设备Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID（6位hex）",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/api/devices/{deviceId}/token": {
            "get": {
                "description": "使用设备6位hex ID认证（路径参数），无需额外Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "获取设备Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID（6位hex）",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/devices/{deviceId}/update": {
            "post": {
                "description": "仅设备所有者或设备自身可更新；增量更新，仅请求体中出现的字段会被更新，未传字段保持原值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "更新设备信息（增量）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID（6位hex）",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "待更新的设备字段（至少一个）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DeviceUpdateParameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ent.Device"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "UserAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "common.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "ent.Device": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "deviceId": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "deviceName": {
                    "description": "DeviceName holds the value of the \"device_name\" field.",
                    "type": "string"
                },
                "deviceType": {
                    "description": "DeviceType holds the value of the \"device_type\" field.",
                    "type": "string"
                },
                "firmwareVersion": {
                    "description": "FirmwareVersion holds the value of the \"firmware_version\" field.",
                    "type": "string"
                },
                "ipAddress": {
                    "description": "IPAddress holds the value of the \"ip_address\" field.",
                    "type": "string"
                },
                "lastActiveTime": {
                    "description": "LastActiveTime holds the value of the \"last_active_time\" field.",
                    "type": "string"
                },
                "location": {
                    "description": "Location holds the value of the \"location\" field.",
                    "type": "string"
                },
                "macAddress": {
                    "description": "MACAddress holds the value of the \"mac_address\" field.",
                    "type": "string"
                },
                "ownerId": {
                    "description": "OwnerID holds the value of the \"owner_id\" field.",
                    "type": "integer"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "service.DeviceParameters": {
            "type": "object",
            "required": [
                "deviceName"
            ],
            "properties": {
                "deviceName": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "string"
                },
                "firmwareVersion": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "macAddress": {
                    "type": "string"
                }
            }
        },
        "service.DeviceRegisterResponse": {
            "type": "object",
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "deviceToken": {
                    "type": "string"
                }
            }
        },
        "service.DeviceUpdateParameters": {
            "type": "object",
            "properties": {
                "deviceName": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "string"
                },
                "firmwareVersion": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "macAddress": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "DeviceAuth": {
            "description": "设备 Token（设备注册后获取）",
            "type": "apiKey",
            "name": "X-Device-Token",
            "in": "header"
        },
        "UserAuth": {
            "description": "Bearer Token（用户 Sa-Token），登录后获取",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}